gl_vbo_dynamicvertex and gl_vbo_dynamicindex now use R_BufferData_Store rather than the much slower path they used before
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@11957 d7cf8633-e32d-0410-b094-e92efae38249

+++ gl_backend.h
@@ -113,7 +113,7 @@ void R_Mesh_ResetTextureState(void);
 void R_Mesh_ClearBindingsForTexture(int texnum);
 
 // renders a mesh
-void R_Mesh_Draw(int firstvertex, int numvertices, int firsttriangle, int numtriangles, const int *element3i, const r_meshbuffer_t *element3i_indexbuffer, size_t element3i_bufferoffset, const unsigned short *element3s, const r_meshbuffer_t *element3s_indexbuffer, size_t element3s_bufferoffset);
+void R_Mesh_Draw(int firstvertex, int numvertices, int firsttriangle, int numtriangles, const int *element3i, const r_meshbuffer_t *element3i_indexbuffer, int element3i_bufferoffset, const unsigned short *element3s, const r_meshbuffer_t *element3s_indexbuffer, int element3s_bufferoffset);
 
 // saves a section of the rendered frame to a .tga or .jpg file
 qboolean SCR_ScreenShot(char *filename, unsigned char *buffer1, unsigned char *buffer2, int x, int y, int width, int height, qboolean flipx, qboolean flipy, qboolean flipdiagonal, qboolean jpeg, qboolean png, qboolean gammacorrect, qboolean keep_alpha);
